4.21*: blogilistan laajennus, step10

+++ controllers/blogs.js
@@ -9,16 +9,17 @@ blogsRouter.get('/', async (_request, response) => {
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
-// eslint-disable-next-line consistent-return
 blogsRouter.post('/', async (request, response, next) => {
   const { body, token } = request
 
   try {
-    const decodedToken = await jwt.verify(token, process.env.SECRET)
+    const decodedToken = await jwt
+      .verify(token, process.env.SECRET)
     if (!decodedToken || !decodedToken.id) {
       return response
         .status(401)
         .json({ error: 'token missing or invalid' })
+        .end()
     }
 
     const user = await User.findById(decodedToken.id)
@@ -34,18 +35,49 @@ blogsRouter.post('/', async (request, response, next) => {
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
-    response.status(201).json(savedBlog.toJSON())
+    return response.status(201).json(savedBlog.toJSON())
   } catch (error) {
-    next(error)
+    return next(error)
   }
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const { token } = request
+
   try {
+    const decodedToken = await jwt
+      .verify(token, process.env.SECRET)
+    if (!decodedToken || !decodedToken.id) {
+      return response
+        .status(401)
+        .json({ error: 'token missing or invalid' })
+        .end()
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response
+        .status(404)
+        .json({ error: 'blog not found' })
+        .end()
+    }
+
+    const tokenUser = blog.user
+    if (JSON.stringify(user) === JSON.stringify(tokenUser)) {
+      return response
+        .status(401)
+        .json({ error: 'wrong authentication' })
+        .end()
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    return response
+      .status(204)
+      .end()
   } catch (error) {
-    next(error)
+    return next(error)
   }
 })
